requirements.txt
PyMySQL==0.7.11
PyYAML==3.12
requests==2.18.4
SQLAlchemy==1.1.14
tornado==4.5.2
PyJWT==1.5.3
ldap3==2.3
